Fix child process not closing on unknown command
The child process never exited, because throw_error didn't exit the
process. Actually exiting the process introduces the difficulty of
getting the error code into an actual throw_error in the host
process; I used pipes for that, and select(2) to poll whether the
child has written an error there.